usb: phy: Initialize qmp phy suspend flag to true
The genpd based USB PHY GDSC is controlled by phy suspend/resume
operations. Not initializing the phy suspend flag to TRUE would
skip the very first set_suspend(0) resume operation, as by default
we would assume that phy is in resume state. This would lead to not
triggering the phy runtime resume op hence not powering ON the
phy GDSC. Fix this by explicitly setting the default value of
suspend flag to true.

Change-Id: Ie5ab809a630152935d2747914f332aac8518c188
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>