drivers: dcvs: memlat: Serialize calls to sysfs apis
With memlat enabled in cpucp, some sysfs calls from
userspace may take more time to complete which can
lead to race conditions if clients try and read a
value before it finished updating. Add locking to
the relevant sysfs apis to prevent this. Also update
sample_ms sysfs api to return error if the message
to cpucp fails as this was a warning previously.

Change-Id: I6c7bd7b2298eba2beca18ad8519f0f6b6d797dd4
Signed-off-by: Amir Vajid <quic_avajid@quicinc.com>