zram: qpace: Prevent BIO leaks in submission error paths
Each BIO submitted to the ZRAM driver in the QPaCE path will have an
associated struct qpace_zram_meta. qpace_zram_meta has a kref counter
for each PAGE_SIZE page in the BIO, because we submit each sub-page to
QPaCE individually (i.e. each page we submit has a single associated
qpace_request_data object).

Inside of qpace_zram_submit_bio(), if there is an error whilst
submitting any page of a BIO to QPaCE, we currently leak that BIO. To
see this, notice that we call kref_init() before we start iterating
over each sub-page in the BIO.  It is this reference that we end up
leaking. Before we pass a page to QPaCE via
qpace_control->requets_submit(), we increase the reference count on
that page. If the submission function succeeds, the rest of the ZRAM
/ QPaCE pipeline is responsible for handling the individual page. If
the submission is unsuccessful, we are responsible for releasing that
reference, which happens inside of zram_qpace_req_err_handler(), which
we do call.  However, we return early, meaning that we will not call
the kref_put() call at the end of qpace_zram_submit_bio(), which
causes us to leak a reference.  Fix this.

Change-Id: I9d6c2216e9ffe7ca60050d3010135da2d874e3e4
Signed-off-by: Chris Goldsworthy <quic_cgoldswo@quicinc.com>